+++ src/utils.ts
@@ -625,10 +625,18 @@ export function arrayTypeAndSizeStr(arrayTypeName: string): [string, Array<strin
  */
 export function arrayTypeAndSize(arrayTypeName: string): [string, Array<number>] {
   const [elemTypeName, arraySizes] = arrayTypeAndSizeStr(arrayTypeName);
-  return [elemTypeName, arraySizes.map(size => parseInt(size))];
+  return [elemTypeName, arraySizes.map(size => {
+    const n = parseInt(size);
+
+    if (isNaN(n)) {
+      throw new Error(`arrayTypeAndSize error type ${arrayTypeName} with sub isNaN`);
+    }
+
+    return n;
+  })];
 }
 
-export function toLiteralArrayType(elemTypeName: string, sizes: Array<number>): string {
+export function toLiteralArrayType(elemTypeName: string, sizes: Array<number | string>): string {
   return [elemTypeName, sizes.map(size => `[${size}]`).join('')].join('');
 }
 
@@ -938,7 +946,13 @@ export function resolveType(alias: AliasEntity[], type: string): string {
 
 
   if (isArrayType(type)) {
-    const [elemTypeName, sizes] = arrayTypeAndSize(type);
+    const [elemTypeName, sizes] = arrayTypeAndSizeStr(type);
+    const elemType = resolveType(alias, elemTypeName);
+
+    if (isArrayType(elemType)) {
+      const [elemTypeName_, sizes_] = arrayTypeAndSizeStr(elemType);
+      return toLiteralArrayType(elemTypeName_, sizes.concat(sizes_));
+    }
     return toLiteralArrayType(resolveType(alias, elemTypeName), sizes);
   }
 
@@ -1092,4 +1106,29 @@ export function isInteger(x: unknown): boolean {
 
 
   return false;
+}
+
+
+export function resolveStaticConst(contract: string, type: string, staticConstInt: Record<string, number>): string {
+
+  if (isArrayType(type)) {
+    const [elemTypeName, arraySizes] = arrayTypeAndSizeStr(type);
+
+    const sizes = arraySizes.map(size => {
+      if (/^(\d)+$/.test(size)) {
+        return parseInt(size);
+      } else {
+        const value = (size.indexOf('.') > 0) ? staticConstInt[size] : staticConstInt[`${contract}.${size}`];
+
+        if (!value) {
+          console.warn(`resolve array sub ${size} fail`);
+          return size;
+        }
+        return value;
+      }
+    });
+
+    return toLiteralArrayType(elemTypeName, sizes);
+  }
+  return type;
 }