+++ test/alias.test.ts
@@ -1,34 +1,29 @@
-import { loadDescription } from './helper';
+import { getContractFilePath, loadDescription } from './helper';
 import { assert, expect } from 'chai';
 import { buildContractClass, buildTypeClasses, buildTypeResolver } from '../src/contract';
 import { Bytes, Int } from '../src/scryptTypes';
-
-const jsonDescr = loadDescription('alias_desc.json');
-
-const AliasContract = buildContractClass(jsonDescr);
+import { compileContract } from '../src/utils';
 
 
 
-const { Male, MaleAAA, Female, Name, Age, Token, Person, Block, Height, Time, Coinbase } = buildTypeClasses(jsonDescr);
-
-let man = new Person({
-  name: new Name("68656c6c6f20776f726c6421"),
-  age: new Age(33),
-  token: new Token(101)
-});
-
-
-
-
-
-const alias = new AliasContract(new Female({
-  name: new Name("68656c6c6f20776f726c6421"),
-  age: new Age(1),
-  token: new Token(101)
-}));
 
 
 describe('Alias type check', () => {
+  const AliasContract = buildContractClass(loadDescription('alias_desc.json'));
+
+  const { Male, MaleAAA, Female, Name, Age, Token, Person, Block, Height, Time, Coinbase } = buildTypeClasses(loadDescription('alias_desc.json'));
+
+  let man = new Person({
+    name: new Name("68656c6c6f20776f726c6421"),
+    age: new Age(33),
+    token: new Token(101)
+  });
+
+  const alias = new AliasContract(new Female({
+    name: new Name("68656c6c6f20776f726c6421"),
+    age: new Age(1),
+    token: new Token(101)
+  }));
 
   it('should success when using MaleAAA', () => {
 
@@ -116,13 +111,22 @@ describe('Alias type check', () => {
   })
 
 
+  it('should success when parameter is array struct have member with alias', () => {
+    let result = alias.isPerson([new Female({
+      name: new Name("68656c6c6f20776f726c6421"),
+      age: new Age(1),
+      token: new Token(101)
+    })]).verify()
+    assert.isTrue(result.success, result.error);
+  })
 
-  describe('Alias buildTypeResolver', () => {
+
+  describe('buildTypeResolver', () => {
 
 
     it('should success when call buildTypeResolver', () => {
-
-      const finalTypeResolver = buildTypeResolver(jsonDescr.alias)
+      const jsondesc = loadDescription('alias_desc.json');
+      const finalTypeResolver = buildTypeResolver(jsondesc.contract, jsondesc.alias, jsondesc.structs, {})
       expect(finalTypeResolver("Age")).to.equal('int')
       expect(finalTypeResolver("Time")).to.equal('int')
       expect(finalTypeResolver("Name")).to.equal('bytes')
@@ -137,6 +141,9 @@ describe('Alias type check', () => {
       expect(finalTypeResolver("Height")).to.equal('int')
       expect(finalTypeResolver("struct Person {}[3]")).to.equal('struct Person {}[3]')
       expect(finalTypeResolver("struct MaleAAA {}[3]")).to.equal('struct Person {}[3]')
+      expect(finalTypeResolver("struct MaleB {}[1]")).to.equal('struct Person {}[1][3]')
+      expect(finalTypeResolver("struct MaleC {}[5]")).to.equal('struct Person {}[5][2][3]')
+
 
       expect(finalTypeResolver("int")).to.equal('int')
       expect(finalTypeResolver("PubKey")).to.equal('PubKey')
@@ -155,9 +162,15 @@ describe('Alias type check', () => {
 
     it('should success when call buildTypeResolver', () => {
 
-      const finalTypeResolver = buildTypeResolver([])
-      expect(finalTypeResolver("Person")).to.equal('struct Person {}')
-      expect(finalTypeResolver("Block")).to.equal('struct Block {}')
+      const finalTypeResolver = buildTypeResolver('', [], [], {})
+
+      expect(() => {
+        finalTypeResolver("Person")
+      }).to.throw("typeResolver with unknown type Person")
+
+      expect(() => {
+        finalTypeResolver("Person[3]")
+      }).to.throw("typeResolver with unknown elemTypeName Person")
 
       expect(finalTypeResolver("int")).to.equal('int')
       expect(finalTypeResolver("PubKey")).to.equal('PubKey')
@@ -174,4 +187,47 @@ describe('Alias type check', () => {
 
     })
   })
+
+  describe('Alias1Contract check', () => {
+
+    const Alias1Contract = buildContractClass(loadDescription('alias1_desc.json'));
+
+    const { Tokens, TokenArray, Token } = buildTypeClasses(loadDescription('alias1_desc.json'));
+
+
+    it('should success when call buildTypeResolver', () => {
+      const jsondesc = loadDescription('alias1_desc.json');
+      const finalTypeResolver = buildTypeResolver(jsondesc.contract, jsondesc.alias, jsondesc.structs, {})
+      expect(finalTypeResolver("Tokens")).to.equal('int[3]')
+      expect(finalTypeResolver("TokenArray")).to.equal('int[1][3]')
+      expect(finalTypeResolver("TokenAA")).to.equal('int[4][5][1][3]')
+      expect(finalTypeResolver("Tokens[1]")).to.equal('int[1][3]')
+      expect(finalTypeResolver("TokenArray[4][5]")).to.equal('int[4][5][1][3]')
+
+    })
+
+    const alias = new Alias1Contract([1, 3, 3], [[2, 1, 3]]);
+
+    it('should success when unlock', () => {
+
+      let result = alias.unlock([1, 3, 3], [[2, 1, 3]]).verify()
+      assert.isTrue(result.success, result.error);
+    })
+  })
+
+
+  describe('VarAsSub check', () => {
+
+    const result = compileContract(getContractFilePath('varassub.scrypt'));
+    it('should success when call buildTypeResolver', () => {
+
+      const finalTypeResolver = buildTypeResolver(result.contract, result.alias, result.structs, result.staticConst)
+      expect(finalTypeResolver("int[1][SUB]")).to.equal('int[1][3]')
+      expect(finalTypeResolver("int[1][VarAsSub.SUB]")).to.equal('int[1][3]')
+
+    })
+
+  })
+
+
 })