+++ test/compilerWrapper.test.ts
@@ -1,11 +1,12 @@
 import { assert, expect } from 'chai';
 import path = require("path");
-import { loadDescription, getContractFilePath, getInvalidContractFilePath, deleteSource } from './helper'
+import { loadDescription, getContractFilePath, getInvalidContractFilePath, excludeMembers } from './helper'
 import { ABIEntityType, CompileResult, desc2CompileResult, compilerVersion, getDefaultScryptc } from '../src/compilerWrapper';
 import { compileContract, getCIScryptc } from '../src/utils';
 import * as minimist from 'minimist';
 import { writeFileSync, readFileSync } from 'fs';
 import { join } from 'path';
+import { buildContractClass, buildTypeClasses } from '../src/contract';
 
 describe('compile()', () => {
   it('compile successfully', () => {
@@ -145,7 +146,8 @@ describe('compile()', () => {
             "line": 1
           }
         ],
-        type: "SemanticError"
+        type: "SemanticError",
+        relatedInformation: []
 
       }])
 
@@ -174,7 +176,8 @@ describe('compile()', () => {
             "line": 1
           }
         ],
-        type: "SemanticError"
+        type: "SemanticError",
+        relatedInformation: []
       }])
     })
 
@@ -200,7 +203,8 @@ describe('compile()', () => {
             "line": 1
           }
         ],
-        type: "SemanticError"
+        type: "SemanticError",
+        relatedInformation: []
       }])
     })
 
@@ -227,7 +231,8 @@ describe('compile()', () => {
             "line": 1
           }
         ],
-        type: "SemanticError"
+        type: "SemanticError",
+        relatedInformation: []
       }])
     })
 
@@ -254,7 +259,8 @@ describe('compile()', () => {
             "line": 1
           }
         ],
-        type: "SemanticError"
+        type: "SemanticError",
+        relatedInformation: []
       }])
     })
 
@@ -281,7 +287,8 @@ describe('compile()', () => {
             "line": 8
           }
         ],
-        type: "SemanticError"
+        type: "SemanticError",
+        relatedInformation: []
       }, {
         filePath: "const.scrypt",
         message: "Expecting a compile time constant",
@@ -295,7 +302,8 @@ describe('compile()', () => {
             "line": 12
           }
         ],
-        type: "SemanticError"
+        type: "SemanticError",
+        relatedInformation: []
       }])
     })
 
@@ -445,84 +453,156 @@ describe('compile()', () => {
   })
 
 
-  describe('output warnings test', () => {
+  describe('output_warnings test', () => {
 
     function warningsTest(warnings) {
-      warnings.forEach(error => {
-        delete error.filePath;
-      });
-      assert.isTrue(warnings.length === 5);
 
-      expect(warnings).to.deep.include.members([
+      assert.isTrue(warnings.length === 5);
+      expect(excludeMembers(warnings, ['filePath'])).to.deep.include.members([
         {
-          type: 'Warning',
-          position: [
+          "type": "Warning",
+          "position": [
             {
-              "column": 17,
-              "line": 15
+              "line": 15,
+              "column": 17
             },
             {
-              "column": 18,
-              "line": 15
+              "line": 15,
+              "column": 18
             }
           ],
-          message: 'Variable `y` shadows existing binding at 11:22:11:23'
+          "message": "Variable `y` shadows existing binding at ",
+          "relatedInformation": [
+            {
+              "position": [
+                {
+                  "line": 11,
+                  "column": 22
+                },
+                {
+                  "line": 11,
+                  "column": 23
+                }
+              ],
+              "message": ""
+            }
+          ]
         },
         {
-          type: 'Warning',
-          position: [
+          "type": "Warning",
+          "position": [
             {
-              "column": 21,
-              "line": 19
+              "line": 19,
+              "column": 21
             },
             {
-              "column": 22,
-              "line": 19
+              "line": 19,
+              "column": 22
             }
           ],
-          message: 'Variable `y` shadows existing binding at 15:17:15:18'
+          "message": "Variable `y` shadows existing binding at ",
+          "relatedInformation": [
+            {
+              "position": [
+                {
+                  "line": 15,
+                  "column": 17
+                },
+                {
+                  "line": 15,
+                  "column": 18
+                }
+              ],
+              "message": ""
+            }
+          ]
         },
         {
-          type: 'Warning',
-          position: [
+          "type": "Warning",
+          "position": [
             {
-              "column": 17,
-              "line": 34
+              "line": 34,
+              "column": 17
             },
             {
-              "column": 18,
-              "line": 34
+              "line": 34,
+              "column": 18
             }
           ],
-          message: 'Variable `i` shadows existing binding at 32:9:36:10'
+          "message": "Variable `i` shadows existing binding at ",
+          "relatedInformation": [
+            {
+              "position": [
+                {
+                  "line": 32,
+                  "column": 9
+                },
+                {
+                  "line": 36,
+                  "column": 10
+                }
+              ],
+              "message": ""
+            }
+          ]
         },
         {
-          type: 'Warning',
-          position: [
+          "type": "Warning",
+          "position": [
             {
-              "column": 17,
-              "line": 44
+              "line": 44,
+              "column": 17
             },
             {
-              "column": 18,
-              "line": 44
+              "line": 44,
+              "column": 18
             }
           ],
-          message: 'Variable `y` shadows existing binding at 41:32:41:33'
+          "message": "Variable `y` shadows existing binding at ",
+          "relatedInformation": [
+            {
+              "position": [
+                {
+                  "line": 41,
+                  "column": 32
+                },
+                {
+                  "line": 41,
+                  "column": 33
+                }
+              ],
+              "message": ""
+            }
+          ]
         },
         {
-          type: 'Warning',
-          position: [
+          "type": "Warning",
+          "position": [
             {
-              "column": 17,
-              "line": 48
+              "line": 48,
+              "column": 17
             },
             {
-              "column": 18,
-              "line": 48
+              "line": 48,
+              "column": 18
             }
           ],
-          message: 'Variable `y` shadows existing binding at 41:32:41:33'
+          "message": "Variable `y` shadows existing binding at ",
+          "relatedInformation": [
+            {
+              "position": [
+                {
+                  "line": 41,
+                  "column": 32
+                },
+                {
+                  "line": 41,
+                  "column": 33
+                }
+              ],
+              "message": ""
+            }
+          ]
         }
       ])
 
@@ -539,10 +619,9 @@ describe('compile()', () => {
 
       const result = compileContract(getInvalidContractFilePath('varshadow.scrypt'));
       warningsTest(result.warnings);
-      result.errors.forEach(error => {
-        delete error.filePath;
-      });
-      expect(result.errors).to.deep.include.members([
+
+
+      expect(excludeMembers(result.errors, ['filePath'])).to.deep.include.members([
         {
           type: 'SemanticError',
           position: [
@@ -555,7 +634,8 @@ describe('compile()', () => {
               "line": 13
             }
           ],
-          message: "Couldn't match expected type 'int' with actual type 'bytes'"
+          message: "Couldn't match expected type 'int' with actual type 'bytes'",
+          relatedInformation: []
         }
       ])
     })
@@ -564,57 +644,96 @@ describe('compile()', () => {
     it('warnings and errors should be right', () => {
 
       const result = compileContract(getInvalidContractFilePath('varshadow1.scrypt'));
-      result.warnings.forEach(warning => {
-        delete warning.filePath;
-      });
-      result.errors.forEach(error => {
-        delete error.filePath;
-      });
 
-      expect(result.warnings).to.deep.include.members([
+      expect(excludeMembers(result.warnings, ['filePath'])).to.deep.include.members([
         {
-          type: 'Warning',
-          position: [
+          "type": "Warning",
+          "position": [
             {
-              "column": 17,
-              "line": 5
+              "line": 5,
+              "column": 17
             },
             {
-              "column": 18,
-              "line": 5
+              "line": 5,
+              "column": 18
             }
           ],
-          message: 'Variable `x` shadows existing binding at 2:30:2:31'
-        }]
-      )
-      expect(result.errors).to.deep.include.members([
-        {
-          type: 'SemanticError',
-          position: [
+          "message": "Variable `x` shadows existing binding at ",
+          "relatedInformation": [
             {
-              "column": 13,
-              "line": 3
+              "position": [
+                {
+                  "line": 2,
+                  "column": 30
+                },
+                {
+                  "line": 2,
+                  "column": 31
+                }
+              ],
+              "message": ""
+            }
+          ]
+        }
+      ])
+      expect(excludeMembers(result.errors, ['filePath'])).to.deep.include.members([
+        {
+          "type": "SemanticError",
+          "position": [
+            {
+              "line": 3,
+              "column": 13
             },
             {
-              "column": 14,
-              "line": 3
+              "line": 3,
+              "column": 14
             }
           ],
-          message: "Symbol `x` already defined at 2:30:2:31"
+          "message": "Symbol `x` already defined at ",
+          "relatedInformation": [
+            {
+              "position": [
+                {
+                  "line": 2,
+                  "column": 30
+                },
+                {
+                  "line": 2,
+                  "column": 31
+                }
+              ],
+              "message": ""
+            }
+          ]
         },
         {
-          type: 'SemanticError',
-          position: [
+          "type": "SemanticError",
+          "position": [
             {
-              "column": 13,
-              "line": 7
+              "line": 7,
+              "column": 13
             },
             {
-              "column": 14,
-              "line": 7
+              "line": 7,
+              "column": 14
             }
           ],
-          message: "Symbol `x` already defined at 2:30:2:31"
+          "message": "Symbol `x` already defined at ",
+          "relatedInformation": [
+            {
+              "position": [
+                {
+                  "line": 2,
+                  "column": 30
+                },
+                {
+                  "line": 2,
+                  "column": 31
+                }
+              ],
+              "message": ""
+            }
+          ]
         }
       ])
     })
@@ -640,9 +759,113 @@ describe('compile()', () => {
 
   it('check foo ast', () => {
     const result = compileContract(getContractFilePath('foo.scrypt'));
-    deleteSource(result.ast)
+    excludeMembers(result.ast, ['source']);
     //writeFileSync(join(__dirname, './fixture/ast/foo.ast.json'), JSON.stringify(result.ast, null, 4));
     const content = readFileSync(join(__dirname, './fixture/ast/foo.ast.json')).toString();
     expect(JSON.parse(JSON.stringify(result.ast))).to.deep.equal(JSON.parse(content));
   })
+
+
+  it('test_ctc_as_parameter_sub', () => {
+    const result = compileContract(getContractFilePath('ctc.scrypt'));
+    expect(result.abi).to.deep.equal([
+      {
+        "type": "function",
+        "name": "unlock",
+        "index": 0,
+        "params": [
+          {
+            "name": "st1",
+            "type": "St1"
+          },
+          {
+            "name": "st2",
+            "type": "St2"
+          },
+          {
+            "name": "a",
+            "type": "St1[2]"
+          },
+          {
+            "name": "b",
+            "type": "St1[3][2]"
+          },
+          {
+            "name": "c",
+            "type": "int[3]"
+          }
+        ]
+      },
+      {
+        "type": "constructor",
+        "params": [
+          {
+            "name": "st1",
+            "type": "St1"
+          },
+          {
+            "name": "st2",
+            "type": "St2"
+          },
+          {
+            "name": "a",
+            "type": "St1[2]"
+          },
+          {
+            "name": "b",
+            "type": "St1[3][2]"
+          },
+          {
+            "name": "c",
+            "type": "int[3]"
+          }
+        ]
+      }
+    ]);
+
+    expect(result.structs).to.deep.equal([
+      {
+        "name": "St1",
+        "params": [
+          {
+            "name": "x",
+            "type": "int[3]"
+          }
+        ]
+      },
+      {
+        "name": "St2",
+        "params": [
+          {
+            "name": "st1s",
+            "type": "St1[2]"
+          }
+        ]
+      }
+    ])
+
+    expect(result.alias).to.deep.equal([
+      {
+        "name": "St1Array",
+        "type": "St1[2]"
+      }
+    ])
+
+
+
+    const CTCContract = buildContractClass(result);
+
+    const { St1, St2 } = buildTypeClasses(result);
+
+    let st1 = new St1({ x: [1, 3, 45] });
+
+    let st2 = new St2({ st1s: [st1, st1] });
+
+    const ctc = new CTCContract(st1, st2, [st1, st1], [[st1, st1], [st1, st1], [st1, st1]], [1, 3, 3]);
+
+    let verify_result = ctc.unlock(st1, st2, [st1, st1], [[st1, st1], [st1, st1], [st1, st1]], [1, 3, 3]).verify()
+
+    assert.isTrue(verify_result.success, "unlock CTCContract failed")
+  })
+
 })